requirements.txt
allure-pytest==2.8.6
allure-python-commons==2.8.6
atomicwrites==1.4.1
attrs==23.1.0
certifi==2023.7.22
chardet==3.0.4
colorama==0.4.6
crypto==1.4.1
exceptiongroup==1.1.3
idna==2.7
iniconfig==2.0.0
more-itertools==10.1.0
Naked==0.1.32
packaging==23.1
pluggy==0.13.1
py==1.11.0
pycryptodome==3.18.0
PyMySQL==0.9.3
pytest==7.4.0
PyYAML==5.3.1
requests==2.20.0
shellescape==3.8.1
six==1.16.0
tomli==2.0.1
urllib3==1.24.3
wcwidth==0.2.6
